기본 기능 완성

+++ boards/views.py
@@ -12,38 +12,24 @@ def goods_list(request):
     return render(request, 'boards/goods_list.html', context)
 
 def article_detail(request, id):
-    # article = Article.objects.get(id=id)
-    # 위 코드는 모델에서 데이터를 가져오는 건데 예외처리를 곁들이자면 아래와 같다.
-    # 인자로는 모델 클래스, 조건 순으로 넣으면 된다.
-    # 이렇게 하면 위와 동일하게 모델에서 objects.get()을 시도하나
-    # (조건 등으로 인해) 찾은 것이 없다면 404 오류를 발생시킨다.
-    # 비슷한 기능을 하는 함수로 하나를 찾는 get() 대신 여러 개를 찾는 filter()를 사용한
-    # get_list_or_404()가 있다.
+    # 상품의 id로 상품을 찾는다. 없으면 404 오류
     good = get_object_or_404(Goods, id=id)
     
     context = {"good": good}
     return render(request, "boards/good_detail.html", context)
 
 # @login_required 데코레이터로 접근을 제한하고 로그인 페이지를 바로 띄울 수 있다.
-# 이때 로그인 페이지 다음에 유저가 가려고 한 페이지를 django가 기억해 주며 쿼리스트링에 next로 지정되어 있다.
-# 이는 별도 처리를 해주지 않으면 기존의 login() 함수에 지정된 경로로 이동하게 된다.
 @login_required
 def create(request):
-    # title = request.POST.get("title")
-    # content = request.POST.get("content")
-
-    # # 새로운 Article 저장
-    # article = Article.objects.create(title=title, content=content)
-    # return redirect("article_detail", article.id)
-
     if request.method == "POST":
-        # ModelForm을 사용할 경우
+        
         form = GoodsForm(request.POST, request.FILES)  # 데이터가 바인딩된 폼을 만든다.
         if form.is_valid():  # form에 있는 데이터들이 전부 유효한지를 확인
-            
-            # ArticleForm은 단순 form이 아닌 ModelForm이므로 Model을 알고 있다.
-            # 따라서 해당 form 객체의 save() 메서드로 별도의 manager 호출 없이 직접 데이터에 저장하는 것까지 가능하다.
-            article = form.save()
+
+            # 작성자 필드를 작성해야 하므로 바로 commit을 하진 않음.
+            article = form.save(commit=False)
+            article.author = request.user
+            article.save()
             
             # 다른 곳으로 리다이렉트를 하자.
             return redirect("boards:good_detail", article.id)
@@ -59,6 +45,8 @@ def update(request, id):
 
     # ModelForm 사용
     good = get_object_or_404(Goods, id=id)
+    if good.author != request.user:
+        return redirect("boards:good_detail", good.id)
 
     if request.method == "POST":
         
@@ -78,8 +66,23 @@ def update(request, id):
 @require_POST
 def delete(request, id):
     # request에는 기본적으로 로그인한 유저의 정보가 들어 있다.
-    # 이것도 django에서 기본적으로 지원하는 것이다...
-    if request.user.is_authenticated:  # 이 조건을 사용해섯 view 내에 분기 처리를 해볼 수도 있겠다.
+    # 이것도 django에서 기본적으로 지원하는 것이다.
+    if request.user.is_authenticated:  # 이 조건을 사용해서 view 내에 분기 처리를 해볼 수도 있겠다.
         good = get_object_or_404(Goods, id=id)
-        good.delete()
-    return redirect("boards:boards")
+        if good.author == request.user:
+            good.delete()
+            return redirect("boards:boards")
+    return redirect("boards:good_detail")
+
+@require_POST
+def like(request, id):
+    if request.user.is_authenticated:
+        good = get_object_or_404(Goods, id=id)
+        if good.like_users.filter(id=request.user.id).exists():
+            good.like_users.remove(request.user)  # 좋아요 취소
+        else:
+            good.like_users.add(request.user)  # 좋아요 등록
+
+        # form으로 데이터 제출 후 현재 페이지에 그대로 머물기 위해서 추가한 코드
+        return redirect(request.META.get('HTTP_REFERER', 'redirect_if_referer_not_found'))
+    return redirect("accounts:login")
